requirements.txt
Flask==2.0.1
folium==0.12.1.post1
geopandas==0.10.2
pandas==1.1.3
python-dotenv==0.21.1
pytz==2020.1
PyYAML==6.0
redis==4.3.4
requests==2.25.0
rq==1.11.0
us==2.0.2
